requirements.txt
asgiref==3.2.10
boto3==1.14.45
botocore==1.17.45
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-multiselectfield==0.1.12
django-pagination==1.0.7
django-storages==1.9.1
docutils==0.15.2
Faker==4.1.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sightengine==1.4.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
